Run bundle check with the --path option when Bundler.settings[: path] is set
Appraisal does not install gems in the Appraisal file to the --path directory when the gems are already installed to the system directory.

In this commit, I add the --path options to bundle check when  Bundler installs gems in the --path directory so that it searches for gems in that directory.

I'm not sure that my approach is correct, but in my environment, it works well.

=======
rbenv 0.4.0-89-g14bc162
ruby 2.1.0
bundler 1.9.4
=======