Run ``` bundle check ``` with ``` BUNDLE_DISABLE_SHARED_GEMS ```  set to '1' if ``` BUNDLE_PATH ``` is specified
Appraisal does not install gems in the Appraisals file to the ``` BUNDLE_PATH ``` directory if the gems are already installed to another directory.

This commit sets ``` BUNDLE_DISABLE_SHARED_GEMS ``` to '1' and run ``` bundle check ``` if  ``` BUNDLE_PATH ``` is defined in .bundle/config so that Bunlder actually installs gems to the appropriate directory.

I'm not sure that my approach is correct, but in my environment, it looks like working well.

=======
rbenv 0.4.0-89-g14bc162
ruby 2.1.0
bundler 1.9.4
=======